warnings--

+++ src/actuators/act_mgr.cxx
@@ -75,7 +75,7 @@ void Actuator_init() {
     // traverse configured modules
     pyPropertyNode group_node = pyGetNode("/config/actuators", true);
     vector<string> children = group_node.getChildren();
-    printf("Found %ld actuator sections\n", children.size());
+    printf("Found %d actuator sections\n", (int)children.size());
     for ( unsigned int i = 0; i < children.size(); i++ ) {
 	pyPropertyNode section = group_node.getChild(children[i].c_str());
 	sections.push_back(section);
